fix(cute_dsl): address third round of code review comments for SM100 blk64 BSA CuTe-DSL kernel
- cute_dsl_utils: serialize constexpr_tvm_ffi_converter_patched's save/patch/
  restore of the process-global TVM-FFI converter behind a threading.RLock,
  so overlapping cute.compile() calls from different threads can no longer
  interleave and leak a stale wrapper into the global slot; verified via a
  deterministic repro of the leak against the old unlocked logic plus an
  8-thread/30-iteration stress test against the fixed version
- dispatch_helpers: resolve_sm100_blk64_split_workspace now re-validates a
  cached split-KV resolution against the *current* CUDA allocator budget on
  every call instead of trusting it for the process lifetime, so a resolution
  made under since-changed memory pressure can no longer silently outlive its
  validity; re-validation is two-tier (cheap free_bytes-only check first,
  falling back to the ~20x costlier memory_reserved()/memory_allocated()
  query only when needed) to keep the common (stable-memory) path cheap
- dispatch_helpers: _blk64_split_workspace_bytes now takes the combine
  kernel's actual output_dtype instead of assuming q.dtype, fixing an
  undercount of the final-output buffer for Sage-FP8 (fp8 input, bf16
  output); threaded through resolve_sm100_blk64_split_workspace and its one
  call site in bsa_attn_sm100_blk64
- dispatch_helpers: sm100_blk64_auto_kv_splits restores the total_q_tiles>=512
  early-return gate present in the upstream Block-Sparse-Attention source
  this module was adapted from but dropped during the port; separately,
  sm100_blk64_kv_splits_from_count's 256/450/900 tiers are retuned to a
  single ~3500 threshold after B200 measurements showed the old tiers
  regressed 1.1x-2.8x vs kv_splits=1 across multiple Q-tile-count shapes,
  only paying off once active KV blocks reached roughly 3500+ (measured
  4-6x speedup there)
- sparse.py: fix kv_splits docstring -- kv_splits=None disables splitting
  (maps to 1), it is not equivalent to "auto"
- test_vsa_block_sparse.py: update the kv_splits threshold unit test and
  test_vsa_blk64_auto_kv_splits_large_topk's shape/assertions for the
  retuned 3500 threshold, keeping end-to-end coverage of the split-KV/
  combine path

Verified: full blk64 suite (tests/attention/test_vsa_block_sparse.py -k
blk64) passes -- 21 passed, 1 skipped (perf-only).

+++ flashinfer/cute_dsl/sparse/bsa_attn_sm100_blk64.py
@@ -263,7 +263,11 @@ def bsa_attn_sm100_blk64_fwd(
             q_bhsd, q2k_block_index, uniform_block_sparse_num
         )
     kv_splits_i = resolve_sm100_blk64_split_workspace(
-        q_bhsd, head_dim_v, kv_splits_i, allow_fallback=auto_kv_splits
+        q_bhsd,
+        head_dim_v,
+        kv_splits_i,
+        allow_fallback=auto_kv_splits,
+        output_dtype=output_dtype,
     )
     allow_empty_block_nums = kv_splits_i > 1
     if use_clc is None:
